Assignment-2/requirements.txt
pytesseract
streamlit
pdf2image
google-generativeai